Rewrite PatientDeleteSpec to conform to standard
